requirements.txt
appnope==0.1.0
bleach==2.1.1
boto==2.48.0
bz2file==0.98
certifi==2017.7.27.1
chainer==3.0.0
chardet==3.0.4
cycler==0.10.0
Cython==0.27.2
decorator==4.1.2
entrypoints==0.2.3
filelock==2.0.12
flake8==3.4.1
gensim==3.0.1
h5py==2.7.1
html5lib==1.0b10
idna==2.6
ipdb==0.10.3
ipykernel==4.6.1
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.0.1
jedi==0.11.0
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.3.0
MarkupSafe==1.0
matplotlib==2.1.0
mccabe==0.6.1
mecab-python3==0.7
mistune==0.7.4
mock==2.0.0
nbconvert==5.3.1
nbformat==4.4.0
nose==1.3.7
notebook==5.2.0
numpy==1.13.3
olefile==0.44
pandas==0.21.0
pandocfilters==1.4.2
parso==0.1.0
pbr==3.1.1
pep8==1.7.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.3.0
prompt-toolkit==1.0.15
protobuf==3.4.0
ptyprocess==0.5.2
pycodestyle==2.3.1
pyflakes==1.5.0
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
pyzmq==16.0.2
qtconsole==4.3.1
requests==2.18.4
scikit-learn==0.19.1
scipy==0.19.1
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
smart-open==1.5.3
terminado==0.6
testpath==0.3.1
tornado==4.5.2
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.0.3
word2vec==0.9.2
